// src/client/voice/VoiceOpCodes.ts
export enum VoiceOpCodes {
  IDENTIFY = 0,
  SELECT_PROTOCOL = 1,
  READY = 2,
  HEARTBEAT = 3,
  SELECT_PROTOCOL_ACK = 4,
  SPEAKING = 5,
  HEARTBEAT_ACK = 6,
  RESUME = 7,
  HELLO = 8,
  RESUMED = 9,
  CLIENTS_CONNECT = 11,
  VIDEO = 12,
  CLIENT_DISCONNECT = 13,
  SESSION_UPDATE = 14,
  MEDIA_SINK_WANTS = 15,
  VOICE_BACKEND_VERSION = 16,
  CHANNEL_OPTIONS_UPDATE = 17,
  FLAGS = 18,
  SPEED_TEST = 19,
  PLATFORM = 20,
  DAVE_PREPARE_TRANSITION = 21,
  DAVE_EXECUTE_TRANSITION = 22,
  DAVE_TRANSITION_READY = 23,
  DAVE_PREPARE_EPOCH = 24,
  MLS_INVALID_COMMIT_WELCOME = 31,
}

export enum VoiceOpCodesBinary {
  MLS_EXTERNAL_SENDER = 25,
  MLS_KEY_PACKAGE = 26,
  MLS_PROPOSALS = 27,
  MLS_COMMIT_WELCOME = 28,
  MLS_ANNOUNCE_COMMIT_TRANSITION = 29,
  MLS_WELCOME = 30,
}